requirements.txt
Django==5.2.4
psycopg2-binary==2.9.10
stripe==12.3.0
python-dotenv==1.1.1
Pillow==11.1.0
dj-database-url==2.3.0
whitenoise==6.8.2

# Web Scraping & Data Collection
scrapy==2.11.0
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2
playwright==1.40.0
lxml==4.9.3

# Background Tasks & Scheduling
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.6.0

# Real-time Features
channels==4.0.0
channels-redis==4.1.0
websockets==12.0

# Notifications & Alerts
twilio==8.10.0
django-notifications-hq==1.8.0
sendgrid==6.10.0

# Monitoring & Analytics
sentry-sdk==1.38.0
django-extensions==3.2.3

# API & Data Processing
djangorestframework==3.14.0
pandas==2.1.3
numpy==1.25.2

# Price Monitoring & Comparison
price-parser==0.3.4
fuzzywuzzy==0.18.0
python-levenshtein==0.23.0

# Web Scraping and Automation
scrapy==2.11.0
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2
playwright==1.40.0

# Real-time and Background Tasks
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Real-time Communications
channels==4.0.0
channels-redis==4.1.0

# Notifications and Alerts
twilio==8.10.0
django-notifications-hq==1.8.3

# Task Scheduling and Monitoring
django-q==1.3.9
sentry-sdk==1.38.0

# Data Processing and Analysis
pandas==2.1.3
numpy==1.25.2
python-dateutil==2.8.2

# Email Templates and Communications
django-mail-templated==2.6.5
premailer==3.10.0
